moved time-intense collection load tests to integration tests
